requirements.txt
streamlit
python-dotenv
openai
PyMuPDF
whisper
python-docx
pandas

